moves configure ddp to each backend
