+++ pytorch_lightning/accelerators/ddp_cpu_torchelastic_backend.py
@@ -22,7 +22,9 @@ from pytorch_lightning.utilities import AMPType
 from pytorch_lightning.utilities.distributed import rank_zero_only
 from pytorch_lightning.utilities.seed import seed_everything
 from pytorch_lightning.distributed.dist import LightningDistributed
-
+from pytorch_lightning.overrides.data_parallel import LightningDistributedDataParallel
+from torch.nn.parallel import DistributedDataParallel
+from typing import List
 
 try:
     from hydra.utils import to_absolute_path, get_original_cwd
@@ -159,7 +161,7 @@ class DDPCPUTorchElasticBackend(Accelerator):
         device_ids = self.get_device_ids()
 
         # allow user to configure ddp
-        model = model.configure_ddp(model, device_ids)
+        model = self.configure_ddp(model, device_ids)
 
         # set up training routine
         self.trainer.train_loop.setup_training(model)
@@ -171,3 +173,11 @@ class DDPCPUTorchElasticBackend(Accelerator):
         torch.cuda.empty_cache()
 
         return results
+
+    def configure_ddp(
+        self, model: "LightningModule", device_ids: List[int]
+    ) -> DistributedDataParallel:
+        model = LightningDistributedDataParallel(
+            model, device_ids=device_ids, find_unused_parameters=True
+        )
+        return model
